requirements.txt
anyio==3.6.2
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==22.4.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
faiss-cpu==1.7.3
fastapi==0.94.0
filelock==3.9.0
h11==0.14.0
httplib2==0.14.0
httptools==0.5.0
huggingface-hub==0.13.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
packaging==23.0
pexpect==4.6.0
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
pydantic==1.10.6
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
pytz==2019.3
PyYAML==5.3.1
regex==2022.10.31
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0.post1
sniffio==1.3.0
sos==4.4
ssh-import-id==5.10
starlette==0.26.0.post1
systemd-python==234
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
Twisted==18.9.0
typing-extensions==4.5.0
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.21.0
uvloop==0.17.0
wadllib==1.3.3
watchfiles==0.18.1
websockets==10.4
wsproto==0.15.0
zipp==1.0.0
zope.interface==4.7.1
